# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Waste Bank System API",
    "version": "1.0.0",
    "description": "API documentation for E-Waste Bank System"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "paths": {
    "/api/inference": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Run AI inference on an uploaded image",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inference result data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/estimate-price": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Estimate price based on category and weight via regression service",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string"
                  },
                  "weight": {
                    "type": "number"
                  }
                },
                "required": [
                  "category",
                  "weight"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estimated price result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimateResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/articles": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Retrieve a paginated list of articles",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of articles with total count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Article"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Articles"
        ],
        "summary": "Create a new article",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "title",
                  "content",
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Article created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Retrieve a single article by its ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          "404": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "put": {
        "tags": [
          "Articles"
        ],
        "summary": "Update an existing article",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Article updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "Articles"
        ],
        "summary": "Delete an article by its ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Article deleted"
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Admin login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Admin logout",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/auth/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User signup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signup successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/reports": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Retrieve approved e-waste requests report",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Approved requests list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EWasteRequest"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/requests": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Retrieve a paginated list of e-waste requests (Admin)",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of requests to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of requests to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "List of requests with total count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EWasteRequest"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Submit a new e-waste request",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "weight": {
                    "type": "number"
                  },
                  "location": {
                    "type": "string"
                  },
                  "pickupDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "weight",
                  "location",
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EWasteRequest"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/requests/{id}/approve": {
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Approve an e-waste request (Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EWasteRequest"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/requests/{id}/reject": {
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Reject an e-waste request (Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EWasteRequest"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        }
      }
    },
    "/api/requests/me": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Retrieve current user's e-waste request history",
        "responses": {
          "200": {
            "description": "Array of user's requests"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}